app/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
sentencepiece
scikit-learn
pandas
numpy
Pillow
